// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TournamentStatus = {
  SETUP: 'SETUP',
  DRAWING: 'DRAWING',
  GROUP_STAGE: 'GROUP_STAGE',
  KNOCKOUT: 'KNOCKOUT',
  FINISHED: 'FINISHED'
} as const

export type TournamentStatus = (typeof TournamentStatus)[keyof typeof TournamentStatus]


export const MatchStage = {
  GROUP: 'GROUP',
  ROUND_OF_16: 'ROUND_OF_16',
  QUARTERFINAL: 'QUARTERFINAL',
  SEMIFINAL: 'SEMIFINAL',
  FINAL: 'FINAL',
  THIRD_PLACE: 'THIRD_PLACE'
} as const

export type MatchStage = (typeof MatchStage)[keyof typeof MatchStage]


export const MatchStatus = {
  SCHEDULED: 'SCHEDULED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus]


export const MatchSlotType = {
  GROUP_DIRECT: 'GROUP_DIRECT',
  GROUP_POSITION: 'GROUP_POSITION',
  WINNER: 'WINNER',
  LOSER: 'LOSER'
} as const

export type MatchSlotType = (typeof MatchSlotType)[keyof typeof MatchSlotType]


export const SlotPosition = {
  HOME: 'HOME',
  AWAY: 'AWAY'
} as const

export type SlotPosition = (typeof SlotPosition)[keyof typeof SlotPosition]